+++ lib/sm_sms_campaign_webhook.rb
@@ -5,4 +5,19 @@ require "sm_sms_campaign_webhook/version"
 
 # Namespace for SMS campaign webhook.
 module SmSmsCampaignWebhook
+  # @return [SmSmsCampaignWebhook] self for configuration purposes
+  def self.config(&block)
+    yield self if block
+  end
+
+  # @return [Processable] SMS campaign payload processor used by operations
+  def self.processor
+    @processor ||= DefaultProcessor
+  end
+
+  # @param processor [Processable] Custom SMS campaign payload processor
+  # @see Processable
+  def self.processor=(processor)
+    @processor = processor
+  end
 end